resolved some pr suggestions and copied pckage-lock.json from main
