+++ src/Components/ScenePageErrorHandlingWrapper/Internal/ErrorIllustration/ErrorIllustration.tsx
@@ -68,7 +68,6 @@ const ErrorIllustration: React.FC<ErrorIllustrationProps> = ({
                         {''}
                     </Text>
                 </div>
-
                 <PrimaryButton
                     text={buttonText}
                     onClick={onClickAction}
