+++ src/Components/ScenePageErrorHandlingWrapper/ScenePageErrorHandlingWrapper.types.ts
@@ -3,8 +3,8 @@ import { IComponentError } from '../../Models/Constants';
 export interface ScenePageErrorHandlingWrapperProps {
     errors: Array<IComponentError>;
     primaryClickAction?: {
-        buttonText?: string;
-        OnClickAction?: () => void;
+        buttonText: string;
+        onClick: () => void;
     };
     children?: React.ReactNode;
 }
