+++ src/Models/Services/Utils.ts
@@ -198,6 +198,16 @@ export function measureText(str: string, fontSize: number) {
     );
 }
 
+export function getTimeStamp() {
+    const d = new Date();
+    const seconds = d.getSeconds();
+    const minutes = d.getMinutes();
+    const hours = d.getHours();
+    const date = encodeURIComponent(d.toLocaleDateString().replace(/\//g, '-'));
+    const timeStamp = `${date}_${hours - 12}:${minutes}:${seconds}`;
+    return timeStamp;
+}
+
 export function parseExpression(expression: string, twins: any) {
     let result: any = '';
     try {
